+++ .gitignore
@@ -56,3 +56,6 @@ typings/
 
 # dotenv environment variables file
 .env
+
+# Dist dir
+dist
