+++ package.json
@@ -2,15 +2,20 @@
   "name": "aoc-loader",
   "version": "1.2.0",
   "description": "Loads input data from Advent of Code",
-  "main": "index.js",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
+  "files": [
+    "dist"
+  ],
   "scripts": {
+    "build": "tsc",
+    "prepack": "npm run build",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "author": "Johan Lindell",
   "license": "MIT",
   "dependencies": {
-    "request": "^2.88.0",
-    "request-promise-native": "^1.0.5"
+    "got": "^11.8.0"
   },
   "devDependencies": {
     "@typescript-eslint/eslint-plugin": "^4.8.1",
@@ -19,6 +24,7 @@
     "eslint-config-airbnb-base": "^14.2.1",
     "eslint-config-prettier": "^6.15.0",
     "eslint-plugin-import": "^2.22.1",
-    "prettier": "^2.2.0"
+    "prettier": "^2.2.0",
+    "typescript": "^4.1.2"
   }
 }
