+++ tsconfig.eslint.json
@@ -1,0 +1,4 @@
+{
+  "extends": "./tsconfig.json",
+  "include": ["*.ts", ".eslintrc.js", ".prettierrc.js"]
+}
